// frontend/src/app/core/models/role.ts
export enum Role {
  // eslint-disable-next-line @typescript-eslint/naming-convention
  User = "user",
  // eslint-disable-next-line @typescript-eslint/naming-convention
  ExternalUser = "externaluser",
  // eslint-disable-next-line @typescript-eslint/naming-convention
  Moderator = "moderator",
  // eslint-disable-next-line @typescript-eslint/naming-convention
  Admin = "admin",
  demoUser = "demoUser",
  tester = "tester"
}